requirements.txt
coreapi==2.0.9
coverage==4.2
coveralls==1.1
Django==1.10.3
django-angular==0.8.4
django-rest-swagger==2.1.0
djangorestframework==3.5.2
djangorestframework-jwt==1.8.0
docopt==0.6.2
factory-boy==2.7.0
fake-factory==0.7.2
ipaddress==1.0.17
itypes==1.1.0
openapi-codec==1.1.7
pluggy==0.4.0
psycopg2==2.6.2
py==1.4.31
PyJWT==1.4.2
python-dateutil==2.6.0
requests==2.12.1
selenium==3.0.1
simplejson==3.10.0
six==1.10.0
tox==2.5.0
uritemplate==3.0.0
virtualenv==15.1.0
